requirements.txt
Django
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
